requirements.txt
Flask>=2.2
mysql-connector-python>=8.0
beautifulsoup4>=4.12
openpyxl>=3.1
reportlab>=4.0
apscheduler>=3.9
python-dotenv>=1.0
